refactor: prefer server-only modules
